+++ src/utils.rs
@@ -5,14 +5,14 @@ use crate::{
 
 pub fn get_piece_color(
     board: [[Option<(PieceType, PieceColor)>; 8]; 8],
-    coordinates: [i32; 2],
+    coordinates: [i8; 2],
 ) -> Option<PieceColor> {
     board[coordinates[0] as usize][coordinates[1] as usize].map(|(_, piece_color)| piece_color)
 }
 
 pub fn get_piece_type(
     board: [[Option<(PieceType, PieceColor)>; 8]; 8],
-    coordinates: [i32; 2],
+    coordinates: [i8; 2],
 ) -> Option<PieceType> {
     board[coordinates[0] as usize][coordinates[1] as usize].map(|(piece_type, _)| piece_type)
 }
@@ -25,8 +25,8 @@ pub fn get_opposite_color(color: PieceColor) -> PieceColor {
 }
 
 // method to clean the position array to remove impossible positions
-pub fn cleaned_positions(positions: Vec<Vec<i32>>) -> Vec<Vec<i32>> {
-    let mut cleaned_array: Vec<Vec<i32>> = vec![];
+pub fn cleaned_positions(positions: Vec<Vec<i8>>) -> Vec<Vec<i8>> {
+    let mut cleaned_array: Vec<Vec<i8>> = vec![];
     for position in positions {
         if is_valid([position[0], position[1]]) {
             cleaned_array.push(position);
@@ -38,7 +38,7 @@ pub fn cleaned_positions(positions: Vec<Vec<i32>>) -> Vec<Vec<i32>> {
 // Return true forally cell color; false for enemy
 pub fn is_cell_color_ally(
     board: [[Option<(PieceType, PieceColor)>; 8]; 8],
-    coordinates: [i32; 2],
+    coordinates: [i8; 2],
     color: PieceColor,
 ) -> bool {
     match get_piece_color(board, coordinates) {
@@ -47,13 +47,13 @@ pub fn is_cell_color_ally(
     }
 }
 
-pub fn is_valid(coordinates: [i32; 2]) -> bool {
+pub fn is_valid(coordinates: [i8; 2]) -> bool {
     let (y, x) = (coordinates[0], coordinates[1]);
 
     (0..8).contains(&y) && (0..8).contains(&x)
 }
 
-pub fn is_vec_in_array(array: Vec<Vec<i32>>, element: [i32; 2]) -> bool {
+pub fn is_vec_in_array(array: Vec<Vec<i8>>, element: [i8; 2]) -> bool {
     for position in array {
         if position == element {
             return true;
@@ -66,10 +66,10 @@ pub fn is_vec_in_array(array: Vec<Vec<i32>>, element: [i32; 2]) -> bool {
 pub fn get_all_checked_cells(
     board: [[Option<(PieceType, PieceColor)>; 8]; 8],
     color: PieceColor,
-) -> Vec<Vec<i32>> {
-    let mut check_cells: Vec<Vec<i32>> = vec![];
-    for i in 0..8i32 {
-        for j in 0..8i32 {
+) -> Vec<Vec<i8>> {
+    let mut check_cells: Vec<Vec<i8>> = vec![];
+    for i in 0..8i8 {
+        for j in 0..8i8 {
             if get_piece_color(board, [i, j]) == Some(color) {
                 continue;
             }
@@ -88,7 +88,7 @@ pub fn get_all_checked_cells(
     check_cells
 }
 
-pub fn col_to_letter(col: i32) -> String {
+pub fn col_to_letter(col: i8) -> String {
     match col {
         0 => "a".to_string(),
         1 => "b".to_string(),
@@ -109,11 +109,11 @@ pub fn convert_position_into_notation(position: i32) -> String {
     let to_y: i32 = position / 10 % 10;
     let to_x: i32 = position % 10;
 
-    result += &col_to_letter(from_x);
-    result += &format!("{}", (8 - from_y) % 8).to_string();
+    result += &col_to_letter(from_x as i8);
+    result += &format!("{}", (8 - from_y as i8) % 8).to_string();
     result += "-";
-    result += &col_to_letter(to_x);
-    result += &format!("{}", (8 - to_y) % 8).to_string();
+    result += &col_to_letter(to_x as i8);
+    result += &format!("{}", (8 - to_y as i8) % 8).to_string();
 
     result
 }
